make Config.js a module
