+++ index.php
@@ -1871,7 +1871,7 @@ if(isset($_SERVER['HTTP_USER_AGENT'])) {
 	<?php
 	if ($debug){
 	?>
-		<script src="/resources/js/Utility/Config.js?v=<?=$debugTime?>" type="text/javascript"></script>
+		<script src="/resources/js/Utility/Config.js?v=<?=$debugTime?>" type="module"></script>
 		<script src="/resources/js/Utility/HelperFunctions.js?v=<?=$debugTime?>" type="text/javascript"></script>
 		<script src="/resources/js/Utility/LayerImgDates.js?v=<?=$debugTime?>" type="text/javascript"></script>
 		<script src="/resources/js/Utility/ClosestImages.js?v=<?=$debugTime?>" type="text/javascript"></script>
